Issue #947: extract cli crate
