+++ crates/tau-cli/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "tau-cli"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+clap = { version = "4", features = ["derive", "env"] }
+
+[dependencies.tau-multi-channel]
+path = "../tau-multi-channel"
