+++ crates/tau-cli/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod cli_args;
+pub mod cli_types;
+
+pub use cli_args::Cli;
+pub use cli_types::*;
